# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de voos",
        "version": "0.1"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Lista de voos",
                "description": "Retorna todos os voos.",
                "operationId": "aca8dfb4861192babc7b0bfd64e876f1",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": [
                                                [
                                                    {
                                                        "id": 1,
                                                        "cia": "GOL",
                                                        "fare": "1AF",
                                                        "flightNumber": "G3-1701",
                                                        "origin": "CNF",
                                                        "destination": "BSB",
                                                        "departureDate": "29/01/2021",
                                                        "arrivalDate": "29/01/2021",
                                                        "departureTime": "07:40",
                                                        "arrivalTime": "09:00",
                                                        "classService": 3,
                                                        "price": 50,
                                                        "tax": 36,
                                                        "outbound": 1,
                                                        "inbound": 0,
                                                        "duration": "1:20"
                                                    },
                                                    {
                                                        "id": 2,
                                                        "cia": "AZUL",
                                                        "fare": "1AF",
                                                        "flightNumber": "AD-1111",
                                                        "origin": "CNF",
                                                        "destination": "BSB",
                                                        "departureDate": "29/01/2021",
                                                        "arrivalDate": "29/01/2021",
                                                        "departureTime": "07:40",
                                                        "arrivalTime": "09:00",
                                                        "classService": 3,
                                                        "price": 50,
                                                        "tax": 36,
                                                        "outbound": 1,
                                                        "inbound": 0,
                                                        "duration": "1:20"
                                                    }
                                                ]
                                            ]
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Erro ao realizar operação."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups-available": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Lista os grupos de voos disponíveis",
                "description": "Retorna o grupo de voos disponiveis",
                "operationId": "26c0943a86d99f4ea2d059688733c4d7",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": [
                                                [
                                                    "1AFG1 (Valor Total R$ 200 | idas[1,2,3] | voltas[9,10])",
                                                    "1AFG2 (Valor Total R$ 250 | idas[1,2,3,4,5,6] | voltas[9,10,11])",
                                                    "1AFG4 (Valor Total R$ 270 | idas[7] | voltas[9,10])",
                                                    "1AFG3 (Valor Total R$ 300 | idas[4,5,6,8] | voltas[9,10,11])"
                                                ]
                                            ]
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Erro ao realizar operação."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/lowest-price": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Retorna o grupo de voo de menor preço",
                "description": "Retorna o grupo com o menor preço",
                "operationId": "31849cef2913062be3fc75dcc785fa69",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": [
                                                {
                                                    "id": "1AFG1",
                                                    "type": "1AF",
                                                    "going": "1,2,3",
                                                    "return": "9,10",
                                                    "price": 200
                                                }
                                            ]
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Erro ao realizar operação."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/informations": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Retorna o grupo de voo de menor preço",
                "description": "Retorna todas as informações de voos",
                "operationId": "beab79a355e6448a7f483fd4483cd653",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": [
                                                {
                                                    "fligths": [
                                                        {
                                                            "id": 1,
                                                            "cia": "GOL",
                                                            "fare": "1AF",
                                                            "flightNumber": "G3-1701",
                                                            "origin": "CNF",
                                                            "destination": "BSB",
                                                            "departureDate": "29/01/2021",
                                                            "arrivalDate": "29/01/2021",
                                                            "departureTime": "07:40",
                                                            "arrivalTime": "09:00",
                                                            "classService": 3,
                                                            "price": 50,
                                                            "tax": 36,
                                                            "outbound": 1,
                                                            "inbound": 0,
                                                            "duration": "1:20"
                                                        },
                                                        {
                                                            "id": 2,
                                                            "cia": "AZUL",
                                                            "fare": "1AF",
                                                            "flightNumber": "AD-1111",
                                                            "origin": "CNF",
                                                            "destination": "BSB",
                                                            "departureDate": "29/01/2021",
                                                            "arrivalDate": "29/01/2021",
                                                            "departureTime": "07:40",
                                                            "arrivalTime": "09:00",
                                                            "classService": 3,
                                                            "price": 50,
                                                            "tax": 36,
                                                            "outbound": 1,
                                                            "inbound": 0,
                                                            "duration": "1:20"
                                                        }
                                                    ],
                                                    "totalGroups": 8,
                                                    "totalFlights": 15,
                                                    "cheapestPrice": 200,
                                                    "cheapestGroup": "1AFG1"
                                                }
                                            ]
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Operação realizada com sucesso."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Exemplos de possíveis erros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Erro ao realizar operação."
                                        },
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}